Update index.ts
CSV is already getting converted to SQL. I added funtions to also read XLSX files. There is also a function to handle both csv and xlsv file. 

If there are any changes or modifications to be made, kindly let me know. 

This is a great initiative and very practical. Im glad to be a part of it. Thank You.